Refactor tests and add fixtures
