+++ tests/unit/test_api.py
@@ -1,4 +1,4 @@
-"""Unit tests for API endpoints"""
+"""Unit tests for API endpoints."""
 import pytest
 from httpx import AsyncClient
 
@@ -15,7 +15,14 @@ class TestHealthEndpoints:
 
 
 class TestAuthentication:
-    def test_login(self):
-        """Test login endpoint"""
-        # Test will be implemented
-        pass
+    @pytest.mark.asyncio
+    async def test_login(self):
+        """Test login endpoint."""
+        async with AsyncClient(app=app, base_url="http://test") as client:
+            response = await client.post(
+                "/api/auth/token",
+                json={"username": "admin", "password": "password"},
+            )
+        assert response.status_code == 200
+        data = response.json()
+        assert "access_token" in data
